pyproject.toml
[tool.poetry]
name = "django-test"
version = "0.1.0"
description = ""
authors = ["Vadim <71ooo@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.16"
django = "^4.1.7"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
whitenoise = "^6.4.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.5"
django4-background-tasks = "^1.2.7"
gunicorn = "^20.1.0"
channels = {extras = ["daphne"], version = "^4.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
